Refactor and correct bugs
- Refactor code and project structure
- Correct a bug on the edit todo functionality (when deleting a task while editing, the sorting options did not appear)
