+++ src/components/organisms/TodoList.jsx
@@ -2,48 +2,22 @@
 import React, { useState } from "react";
 import { useSelector, useDispatch } from "react-redux";
 // ---------- Components imports
-import TodoFilter from "./TodoFilter";
-import TodoOrder from "./TodoOrder";
-import TodoItem from "./TodoItem";
-import TodoTextItem from "./TodoTextItem";
+import SortableTodoList from "../molecules/SortableToDoList"
+import TodoFilter from "../molecules/TodoFilter";
+import TodoOrder from "../molecules/TodoOrder";
 import Paper from "@material-ui/core/Paper";
-import List from "@material-ui/core/List";
 // ---------- Layout imports
 import Grid from "@material-ui/core/Grid";
 import Divider from "@material-ui/core/Divider";
-import { SortableContainer, SortableElement } from "react-sortable-hoc";
 // ---------- Utils imports
 import { FILTER_OPTIONS } from "../../utils/filterOptions";
 import arrayMove from "array-move";
 import { moveTodo } from "../../redux_store/todosReducer";
 import { getTotalTodos } from "../../utils/todos";
 // ---------- Styles imports
-import { listTheme } from "../../styles/list_style";
 import { filterTheme } from "../../styles/filter_styles";
 import { useFilterStyles } from "../../styles/filter_styles";
 import { ThemeProvider } from "@material-ui/core/styles";
-
-// ---------- Images imports
-
-const SortableItem = SortableElement(({ value }) => <TodoItem todo={value} />);
-
-const SortableTodoList = SortableContainer(({ todos, isFiltering }) => {
-  return (
-    <ThemeProvider theme={listTheme}>
-      <List component="nav">
-        {todos.map((todo, index) => (
-          <div key={todo.id}>
-            {!isFiltering && !todo.deleted && (
-              <SortableItem index={index} value={todo} />
-            )}
-            {isFiltering && <TodoTextItem todo={todo} />}
-            <Divider />
-          </div>
-        ))}
-      </List>
-    </ThemeProvider>
-  );
-});
 
 const TodoList = () => {
   const filterClasses = useFilterStyles();
@@ -52,7 +26,6 @@ const TodoList = () => {
   const isEditing = useSelector((state) => state.app.isEditing);
   const dispatch = useDispatch();
 
-  const [todosFiltered, setTodosFiltered] = useState([]);
   const [filterOption, setFilterOption] = useState(FILTER_OPTIONS.all);
   const [search, setSearch] = useState("");
